mcp-server/requirements.txt
mcp[cli]>=1.0.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.0
